[CoreEventDispatcher] Renamed.
