+++ src/index.tsx
@@ -19,7 +19,7 @@ import {ThemeProvider} from "@material-ui/core/styles";
 import theme from "./theme";
 import { DndProvider } from 'react-dnd'
 import { HTML5Backend } from 'react-dnd-html5-backend'
-import WebSocketCore from "./core/WebSocketCore";
+import WebSocketCore from "./core/CoreEventDispatcher";
 
 ReactDOM.render(
     <React.StrictMode>
